changes to cli test fixtures for central use with other testing
